🚧 Tested some stuff
